notification services added and handler are changed to 1

+++ src/main/java/com/ecommerce/ecommApp/commons/pojo/customer/Customer.java
@@ -11,10 +11,9 @@ public class Customer {
     @JsonProperty("name")
     private String name;
     @JsonProperty("number")
-    private Integer number;
+    private Long number;
     @JsonProperty("email")
     private String email;
     @JsonProperty("whatsapp")
-    private Integer whatsapp;
-
+    private Long whatsapp;
 }
